fold MQTTHandler to logging_mqtt_handler.py
